# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "Fluxa API"
  version: "1.0.0"
  description: |
    Fluxa is a multi-currency digital asset platform built on Stellar. It provides
    wallets, instant transfers, FX conversions, scheduled recurring payments,
    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.

    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.
  contact:
    name: "Fluxa Engineering"
    email: "engineering@fluxa.io"
    url: "https://github.com/fluxa/fluxa"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: http://localhost:3000
    description: "Local development"
  - url: https://api.fluxa.io
    description: "Production"
...
security:
  - BearerAuth: []
paths:
...
  /v1/fiat/quote:
    post:
      operationId: previewFiat
      summary: "Preview a fiat deposit or withdrawal"
      description: "Get a read-only quote before creating a fiat deposit or withdrawal."
      tags: ["Wallets"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [side, amount, currency, country]
              properties:
                side:
                  type: string
                  enum: [deposit, withdraw]
                amount:
                  type: string
                currency:
                  type: string
                country:
                  type: string
      responses:
        "200":
          description: "Preview quote"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiatQuotePreviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── FX ────────────────────────────────────────────────────────────────────
...
  /v1/webhooks/fiat/{provider}:
    post:
      operationId: handleFiatWebhook
      summary: "Handle a fiat provider webhook"
      description: |
        Receives webhook callbacks from fiat payment providers (e.g. Flutterwave).
        This is an internal endpoint called by the provider, not by the tenant.
      tags: ["System"]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: "Provider name (e.g. flutterwave)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Provider-specific webhook payload"
      responses:
        "200":
          description: "Webhook processed"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        "400":
          description: "Webhook validation failed"

  # ── Fees ──────────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Pass a tenant API key as `Bearer <key>` in the `Authorization` header.
        Keys are prefixed with `sk_live_` (production) or `sk_test_` (sandbox).
    AdminAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Pass an admin API key as `Bearer <key>`. Admin keys are prefixed with
        `sk_admin_`.

  # ── Parameters ──────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "BAD_REQUEST"
              message: "invalid request body"
    Unauthorized:
      description: "Unauthorized — missing or invalid API key"
      content:
        application/json:
          schema:
            type: object
          example:
            error: "missing or invalid authorization header"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: ["error"]
      properties:
        error:
          type: object
          required: ["code", "message"]
          properties:
            code:
              type: string
              enum:
                - "BAD_REQUEST"
                - "NOT_FOUND"
                - "UNAUTHORIZED"
                - "FORBIDDEN"
                - "INTERNAL_ERROR"
                - "QUOTE_EXPIRED"
                - "QUOTE_ALREADY_USED"
              description: "Machine-readable error code"
            message:
              type: string
              description: "Human-readable error message"

    # ── Health ────────────────────────────────────────────────────────────
...
    FiatQuotePreviewResponse:
      type: object
      properties:
        provider:
          type: string
        fiat_amount:
          type: string
        fiat_currency:
          type: string
        usdc_amount:
          type: string
        rate:
          type: string
        fee:
          type: string
        min_limit:
          type: string
        max_limit:
          type: string
        expires_at:
          type: string
          format: date-time
...